requirements.txt
asgiref==3.10.0
certifi==2025.10.5
distlib==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
packaging==25.0
pillow==12.0.0
pipenv==2025.0.4
platformdirs==4.5.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
django-storages==1.14.4
boto3==1.34.113
setuptools==80.9.0
sqlparse==0.5.3
virtualenv==20.35.4
